Sort `.gitmodules` as part of packaging script
