+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@aws-sdk/client-s3": "3.454.0",
     "ajv": "8.12.0",
+    "git-submodule-js": "1.0.4",
     "strip-json-comments": "5.0.1",
     "toml": "3.0.0",
     "tree-sitter-cli": "0.21.0-pre-release-1"
